BE-681 Add test case for etcdraft

+++ app/platform/fabric/e2e-test/steps/config_util.py
@@ -7,6 +7,7 @@
 
 import subprocess
 import os
+import re
 import sys
 from shutil import copyfile
 import uuid
@@ -30,9 +31,23 @@ Organizations:
         Name: {orgName}
         ID: {orgMSP}
         MSPDir: ./peerOrganizations/{orgMSP}/peers/peer0.{orgMSP}/msp
+        AdminPrincipal: Role.ADMIN
         AnchorPeers:
             - Host: peer0.{orgMSP}
-              Port: 7051 '''
+              Port: 7051
+        Policies:
+            Readers:
+                Type: Signature
+                Rule: OR('{orgMSP}.member')
+            Writers:
+                Type: Signature
+                Rule: OR('{orgMSP}.member')
+            Admins:
+                Type: Signature
+                Rule: OR('{orgMSP}.admin')
+            Endorsement:
+                Type: Signature
+                Rule: OR('{orgMSP}.member') '''
 
 ORDERER_STR = '''
 OrdererOrgs:
@@ -47,8 +62,15 @@ PEER_ORG_STR = '''
   - Name: {name}
     Domain: {domain}
     EnableNodeOUs: {ouEnable}
+    CA:
+      Hostname: ca
+      SANS:
+        - 0.0.0.0
+        - localhost
     Template:
       Count: {numPeers}
+      SANS:
+        - 0.0.0.0
     Users:
       Count: {numUsers}
 '''
@@ -61,13 +83,15 @@ def updateEnviron(context):
 
 def makeProjectConfigDir(context, returnContext=False):
     # Save all the files to a specific directory for the test
-    if not hasattr(context, "projectName") and not hasattr(context, "composition"):
+    if hasattr(context, "projectName"):
+        print("test {} already exists!!!".format(context.projectName))
+    elif not hasattr(context, "projectName") and not hasattr(context, "composition"):
         context.projectName = str(uuid.uuid1()).replace('-','')
     elif hasattr(context, "composition"):
         context.projectName = context.composition.projectName
 
     testConfigs = "configs/%s" % context.projectName
-    if not os.path.isdir(testConfigs):
+    if not os.path.exists(testConfigs) and not os.path.isdir(testConfigs):
         os.mkdir(testConfigs)
     if returnContext:
         return testConfigs, context
@@ -116,7 +140,7 @@ def setCAConfig(context):
 
 def setupConfigsForCA(context, channelID):
     testConfigs, context = makeProjectConfigDir(context, returnContext=True)
-    print("testConfigs: {0}".format(testConfigs))
+    print("ca testConfigs: {0}".format(testConfigs))
 
     configFile = "configtx_fca.yaml"
     if os.path.isfile("configs/%s.yaml" % channelID):
@@ -125,7 +149,6 @@ def setupConfigsForCA(context, channelID):
     copyfile("configs/%s" % configFile, "%s/configtx.yaml" % testConfigs)
 
     orgDirs = getOrgs(context)
-    print("Org Dirs: {}".format(orgDirs))
 
     for orgDir in orgDirs:
         copyfile("{0}/configtx.yaml".format(testConfigs),
@@ -151,14 +174,15 @@ def certificateSetupForCA(context):
                      "{0}/{1}/msp/cacerts/ca.{1}-cert.pem".format(testConfigs, orgDir))
 
 def setupConfigs(context, channelID):
-    testConfigs = makeProjectConfigDir(context)
-    print("testConfigs: {0}".format(testConfigs))
+    testConfigs, context = makeProjectConfigDir(context, returnContext=True)
+    print("reg testConfigs: {0}".format(testConfigs))
 
     configFile = "configtx.yaml"
     if os.path.isfile("configs/%s.yaml" % channelID):
         configFile = "%s.yaml" % channelID
 
-    copyfile("configs/%s" % configFile, "%s/configtx.yaml" % testConfigs)
+    if not os.path.exists("{}/configtx.yaml".format(testConfigs)):
+        copyfile("configs/%s" % configFile, "%s/configtx.yaml" % testConfigs)
 
     # Copy config to orderer org structures
     for orgDir in os.listdir("./{0}/ordererOrganizations".format(testConfigs)):
@@ -182,7 +206,6 @@ def inspectOrdererConfig(context, filename, channelID):
                    "-configPath", ".",
                    "-channelID", channelID]
         return subprocess.check_output(command, cwd=testConfigs, env=updated_env)
-        #return subprocess.check_output(command, env=updated_env)
     except:
         print("Unable to inspect orderer config data: {0}".format(sys.exc_info()[1]))
 
@@ -194,7 +217,6 @@ def inspectChannelConfig(context, filename, channelID):
                    "-configPath", ".",
                    "-channelID", channelID]
         return subprocess.check_output(command, cwd=testConfigs, env=updated_env)
-        #return subprocess.check_output(command, env=updated_env)
     except:
         print("Unable to inspect channel config data: {0}".format(sys.exc_info()[1]))
 
@@ -215,12 +237,12 @@ def generateOrdererConfig(context, channelID, ordererProfile, block):
     testConfigs = makeProjectConfigDir(context)
     updated_env = updateEnviron(context)
     try:
-        command = ["configtxgen", "-profile", ordererProfile,
+        command = ["configtxgen",
+                   "-profile", ordererProfile,
                    "-outputBlock", block,
                    "-configPath", ".",
                    "-channelID", channelID]
         subprocess.check_call(command, cwd=testConfigs, env=updated_env)
-        #subprocess.check_call(command, env=updated_env)
     except:
         print("Unable to generate orderer config data: {0}".format(sys.exc_info()[1]))
 
@@ -228,12 +250,12 @@ def generateChannelConfig(channelID, profile, context):
     testConfigs = makeProjectConfigDir(context)
     updated_env = updateEnviron(context)
     try:
-        command = ["configtxgen", "-profile", profile,
+        command = ["configtxgen",
+                   "-profile", profile,
                    "-outputCreateChannelTx", "%s.tx" % channelID,
                    "-configPath", ".",
                    "-channelID", channelID]
         subprocess.check_call(command, cwd=testConfigs, env=updated_env)
-        #subprocess.check_call(command, env=updated_env)
     except:
         print("Unable to generate channel config data: {0}".format(sys.exc_info()[1]))
 
@@ -251,18 +273,19 @@ def generateChannelAnchorConfig(channelID, profile, context):
     updated_env = updateEnviron(context)
     for org in os.listdir("./{0}/peerOrganizations".format(testConfigs)):
         try:
-            command = ["configtxgen", "-profile", profile,
+            command = ["configtxgen",
+                       "-profile", profile,
                        "-outputAnchorPeersUpdate", "{0}{1}Anchor.tx".format(org, channelID),
                        "-channelID", channelID,
                        "-configPath", testConfigs,
                        "-asOrg", org.title().replace('.', '')]
-            #subprocess.check_call(command, cwd=testConfigs, env=updated_env)
             subprocess.check_call(command, env=updated_env)
         except:
             print("Unable to generate channel anchor config data: {0}".format(sys.exc_info()[1]))
 
 def generateCrypto(context, cryptoLoc="./configs/crypto.yaml"):
     testConfigs = makeProjectConfigDir(context)
+
     updated_env = updateEnviron(context)
     try:
         subprocess.check_call(["cryptogen", "generate",
@@ -333,9 +356,11 @@ def mspandtlsCheck(path, tlsExist):
 
 def fileExistWithExtension(path, message, fileExt=''):
     for root, dirnames, filenames in os.walk(path):
-        assert len(filenames) > 0, "{0}: len: {1}".format(message, len(filenames))
+        assert len(filenames) > 0, "{0}: [{1}] len: {2}".format(message, path, len(filenames))
         fileCount = [filename.endswith(fileExt) for filename in filenames]
         assert fileCount.count(True) >= 1
+        if "tls" in dirnames:
+            break
 
 def rolebasedCertificate(path):
     adminpath = path + "admincerts/"
@@ -344,8 +369,8 @@ def rolebasedCertificate(path):
     capath = path + "cacerts/"
     fileExistWithExtension(capath, "There is not .pem cert in {0}.".format(capath), '.pem')
 
-    signcertspath = path + "signcerts/"
-    fileExistWithExtension(signcertspath, "There is not .pem cert in {0}.".format(signcertspath), '.pem')
+#    signcertspath = path + "signcerts/"
+#    fileExistWithExtension(signcertspath, "There is not .pem cert in {0}.".format(signcertspath), '.pem')
 
     tlscertspath = path + "tlscerts/"
     fileExistWithExtension(tlscertspath, "There is not .pem cert in {0}.".format(tlscertspath), '.pem')
@@ -370,7 +395,36 @@ def buildConfigtx(testConfigs, orgName, mspID):
     with open("{}/configtx.yaml".format(testConfigs), "w") as fd:
         fd.write(configtx)
 
-def getNewOrg(context, mspID):
+def changeCapabilities(context, capType, capVersion):
+    testConfigs, context = makeProjectConfigDir(context, returnContext=True)
+    if os.path.exists("{}/configtx.yaml".format(testConfigs)):
+        copyfile("{}/configtx.yaml".format(testConfigs), "{}/orig_configtx.yaml".format(testConfigs))
+    else:
+        copyfile("configs/configtx.yaml", "{}/configtx.yaml".format(testConfigs))
+
+    with open("{}/configtx.yaml".format(testConfigs), "r") as fd:
+        config = fd.read()
+
+    upgradeConfig = config.replace("{0}: &{0}Capabilities\n        V1_3: true".format(capType),
+                                   "{0}: &{0}Capabilities\n        {1}: true".format(capType, capVersion))
+    with open("{}/configtx.yaml".format(testConfigs), "w") as fd:
+        fd.write(upgradeConfig)
+    return context
+
+def getCurrentConfig(context, channel):
+    testConfigs, context = makeProjectConfigDir(context, returnContext=True)
+    updated_env = updateEnviron(context)
+    block = "{}.block".format(channel)
+
+    subprocess.check_output(["configtxlator", "proto_decode", "--input", block, "--type", "common.Block", "--output", "configStrCurrentConfig.pb"], cwd=testConfigs , env=updated_env)
+    with open("{0}/configStrCurrentConfig.pb".format(testConfigs), "r") as fd:
+        config = json.loads(fd.read())
+
+    with open("{0}/config.json".format(testConfigs), "w") as fd:
+        fd.write(json.dumps(config["data"]["data"][0]["payload"]["data"]["config"], indent=4))
+    return config
+
+def getNewOrg(context, mspID, channelID):
     testConfigs, context = makeProjectConfigDir(context, returnContext=True)
     updated_env = updateEnviron(context)
 
@@ -378,8 +432,12 @@ def getNewOrg(context, mspID):
     copyfile("{}/configtx.yaml".format(testConfigs), "{}/orig_configtx.yaml".format(testConfigs))
     buildConfigtx(testConfigs, orgName, mspID)
     try:
-        command = ["configtxgen", "-printOrg", orgName]
-        args = subprocess.check_output(command, cwd=testConfigs, env=updated_env)
+        command = ["configtxgen",
+                "-configPath", testConfigs,
+                "-channelID", channelID,
+                "-printOrg",
+                orgName]
+        args = subprocess.check_output(command, env=updated_env)
         print("Result of printOrg: ".format(args))
     except:
         print("Unable to inspect orderer config data: {0}".format(sys.exc_info()[1]))
@@ -393,16 +451,7 @@ def getNewOrg(context, mspID):
 def delNewOrg(context, group, mspID, channel):
     updated_env = updateEnviron(context)
     testConfigs = "./configs/{0}".format(context.projectName)
-    inputFile = "{0}.block".format(channel)
-
-    # configtxlator proto_decode --input config_block.pb --type common.Block | jq .data.data[0].payload.data.config > config.json
-    configStr = subprocess.check_output(["configtxlator", "proto_decode", "--input", inputFile, "--type", "common.Block"], cwd=testConfigs , env=updated_env)
-    config = json.loads(configStr)
-
-    with open("{0}/config.json".format(testConfigs), "w") as fd:
-        fd.write(json.dumps(config["data"]["data"][0]["payload"]["data"]["config"], indent=4))
-
-    # configtxlator proto_encode --input config.json --type common.Config --output config.pb
+    config = getCurrentConfig(context, channel)
     subprocess.check_output(["configtxlator", "proto_encode", "--input", "config.json", "--type", "common.Config", "--output", "config.pb"],
                                         cwd=testConfigs,
                                         env=updated_env)
@@ -457,23 +506,26 @@ def registerWithABAC(context, user):
 def addNewOrg(context, config_update, group, channel):
     updated_env = updateEnviron(context)
     testConfigs = "./configs/{0}".format(context.projectName)
-    inputFile = "{0}.block".format(channel)
-
-    # configtxlator proto_decode --input config_block.pb --type common.Block | jq .data.data[0].payload.data.config > config.json
-    configStr = subprocess.check_output(["configtxlator", "proto_decode", "--input", inputFile, "--type", "common.Block"], cwd=testConfigs , env=updated_env)
-    config = json.loads(configStr)
-
-    with open("{0}/config.json".format(testConfigs), "w") as fd:
-        fd.write(json.dumps(config["data"]["data"][0]["payload"]["data"]["config"], indent=4))
-
-    # configtxlator proto_encode --input config.json --type common.Config --output config.pb
+    config = getCurrentConfig(context, channel)
     subprocess.check_output(["configtxlator", "proto_encode", "--input", "config.json", "--type", "common.Config", "--output", "config.pb"],
                                         cwd=testConfigs,
                                         env=updated_env)
 
     # groups = "Application"
     # config_update = {"Org3ExampleCom": <data>}
+    print(config_update)
     config["data"]["data"][0]["payload"]["data"]["config"]["channel_group"]["groups"][group]["groups"].update(config_update)
+    return config
+
+def updateCapabilityConfig(context, channel, capType, capVers):
+    updated_env = updateEnviron(context)
+    testConfigs = "./configs/{0}".format(context.projectName)
+    config = getCurrentConfig(context, channel)
+    subprocess.check_output(["configtxlator", "proto_encode", "--input", "config.json", "--type", "common.Config", "--output", "config.pb"],
+                                        cwd=testConfigs,
+                                        env=updated_env)
+
+    config['data']['data'][0]['payload']['data']['config']['channel_group']['groups'][capType]['values']['Capabilities']['value']['capabilities'] = {capVers: {}}
     return config
 
 def configUpdate(context, config, group, channel):
@@ -483,7 +535,7 @@ def configUpdate(context, config, group, channel):
     with open("{0}/modified_config.json".format(testConfigs), "w") as fd:
         fd.write(json.dumps(config["data"]["data"][0]["payload"]["data"]["config"], indent=4))
 
-    print("Modified config: {}".format(config["data"]["data"][0]["payload"]["data"]["config"]["channel_group"]["groups"][group]["groups"]))
+    print("Modified config: {}".format(config["data"]["data"][0]["payload"]["data"]["config"]["channel_group"]["groups"][group]["values"]))
 
     # configtxlator proto_encode --input config.json --type common.Config --output config.pb
     subprocess.check_output(["configtxlator", "proto_encode", "--input", "config.json", "--type", "common.Config", "--output", "config.pb"],
@@ -501,10 +553,11 @@ def configUpdate(context, config, group, channel):
                                         env=updated_env)
 
     # configtxlator proto_decode --input update.pb --type common.ConfigUpdate | jq . > org3_update.json
-    configStr = subprocess.check_output(["configtxlator", "proto_decode", "--input", "update.pb", "--type", "common.ConfigUpdate"],
-                                        cwd=testConfigs,
-                                        env=updated_env)
-    config = json.loads(configStr)
+    subprocess.check_output(["configtxlator", "proto_decode", "--input", "update.pb", "--type", "common.ConfigUpdate", "--output", "configStrForUpdate.pb"],
+                            cwd=testConfigs,
+                            env=updated_env)
+    with open("{0}/configStrForUpdate.pb".format(testConfigs), "r") as fd:
+        config = json.loads(fd.read())
 
     # echo '{"payload":{"header":{"channel_header":{"channel_id":"mychannel", "type":2}},"data":{"config_update":'$(cat org3_update.json)'}}}' | jq . > org3_update_in_envelope.json
     updatedconfig = {"payload": {"header": {"channel_header": {"channel_id": channel,
@@ -523,3 +576,26 @@ def configUpdate(context, config, group, channel):
                                         env=updated_env)
 
     return "{0}/update{1}.pb".format(testConfigs, channel)
+
+def generateCollections(context, collectionsFile):
+    template = "./configs/collections.json"
+    pattern = r"github.com/hyperledger/fabric-test/(?P<chaincodePath>[\s\S]*)"
+    result = re.match(pattern, context.chaincode["path"])
+    if result is not None:
+        path = "../{0}".format(result.groupdict()["chaincodePath"])
+        if os.path.exists(template):
+            template = "./{0}/collections.json".format(path)
+
+    structure = {}
+    structure["name"] = context.chaincode["name"]
+    structure["policy"] = str(context.chaincode["policy"].replace(" ", "").replace("\'", "'"))
+    print(structure)
+
+    with open(template, "r") as tfd:
+        collectionConfig = json.loads(tfd.read() % (structure))
+
+    finalConfig = "configs/{0}/{1}".format(context.projectName, collectionsFile)
+    with open(finalConfig, "w") as fd:
+        fd.write(json.dumps(collectionConfig, indent=2))
+
+    return "/var/hyperledger/{}".format(finalConfig)
